MongoProject/src/test/java/com/test/controller/ApiControllerTest.java
package com.test.controller;


import com.fasterxml.jackson.databind.ObjectMapper;
import com.fasterxml.jackson.databind.ObjectWriter;
import com.fasterxml.jackson.databind.SerializationFeature;
import com.mongo.controller.ApiController;
import com.mongo.entity.Address;
import com.mongo.entity.User;
import com.mongo.service.UserService;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;
import org.mockito.InjectMocks;
import org.mockito.MockitoAnnotations;
import org.springframework.boot.test.context.SpringBootTest;
import org.springframework.boot.test.mock.mockito.MockBean;
import org.springframework.context.annotation.Profile;
import org.springframework.http.MediaType;
import org.springframework.test.web.servlet.MockMvc;

import static com.test.Util.createUserDetails;
import static org.junit.jupiter.api.Assertions.assertThrows;
import static org.junit.jupiter.api.Assertions.assertTrue;
import static org.mockito.ArgumentMatchers.any;
import static org.mockito.Mockito.when;
import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;
import static org.springframework.test.web.servlet.setup.MockMvcBuilders.standaloneSetup;

@SpringBootTest
public class ApiControllerTest {

    private MockMvc mockMvc;
    @InjectMocks
    private ApiController apiController = new ApiController();

    @MockBean
    private UserService userService;

    @BeforeEach
    public void setUp() {
        MockitoAnnotations.initMocks(this);
        this.mockMvc = standaloneSetup(apiController).build();
    }

    @Test
    final void createUserTest() throws Exception {
        //Given
        User user = createUserDetails();
        ObjectMapper mapper = new ObjectMapper();
        mapper.configure(SerializationFeature.WRAP_ROOT_VALUE, false);
        ObjectWriter ow = mapper.writer().withDefaultPrettyPrinter();
        String requestJson = ow.writeValueAsString(user);
        //when
        when(userService.addNewUser(any(User.class))).thenReturn(user);

        //then
        this.mockMvc.perform(post("/adduser").contentType(MediaType.APPLICATION_JSON)
                .content(requestJson))
                .andExpect(status().isOk())
                .andExpect(content().contentType(MediaType.APPLICATION_JSON));
    }


    @Test
    final void getUser() throws Exception {
        User user = createUserDetails();
        when(userService.getUserById(3437268)).thenReturn(user);
        this.mockMvc.perform(get("/getuser/3437268")
                .accept(MediaType.APPLICATION_JSON))
                .andExpect(status().isOk())
                .andExpect(content().contentType(MediaType.APPLICATION_JSON))
                .andExpect(jsonPath("$.rollno").value(3437268))
                .andExpect(jsonPath("$.name").value("shane aalam"))
                .andExpect(jsonPath("$.address.pincode").value(21212));

    }

}
